requirements.txt
azure-functions==1.20
openai==1.92.2
python-dotenv==1.1.1
httpx>=0.27
